Refactors Categories -> Category, uses Flatlist from HomeScreen
