+++ TTKDTrainingApp/src/components/VideoThumbnail/VideoThumbnail.js
@@ -4,7 +4,7 @@ import {StyleSheet, Image, View, TouchableOpacity} from 'react-native';
 class VideoThumbnail extends PureComponent {
   render() {
     return (
-      <View>
+      <View style={styles.container}>
         {this.props.videos.map(video => (
           <TouchableOpacity
             key={video.uri}
@@ -24,6 +24,7 @@ class VideoThumbnail extends PureComponent {
 }
 
 const styles = StyleSheet.create({
+  container: {flexDirection: 'row'},
   thumbnail: {
     height: 75,
     width: 75,
